fix: fixed some bad style bugs and cleanup
