--- demo/index.css
@@ -1,3 +1,0 @@
-body {
-  /* background-color: #3c3c3c; */
-}
