+++ demo/index.html
@@ -1,5 +1,13 @@
 <!DOCTYPE html>
 <html>
+  <head>
+    <meta name="viewport" content="initial-scale=1, width=device-width" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
+    />
+    <title>Material Grid Demo</title>
+  </head>
   <body>
     <div id="demo"></div>
     <script src="./index.tsx" type="module"></script>
