+++ src/components/GridColumnHeaders.tsx
@@ -1,24 +1,21 @@
 import type { RowData, Table } from "@tanstack/react-table";
-import { type CSSProperties, type Ref } from "react";
-import GridHeaderCell from "./GridHeaderCell";
-import { GridHeadRoot, GridRow } from "./styled";
+import GridHeadCell from "./GridHeadCell";
+import { GridRow } from "./styled";
+import { GridHeadRoot } from "./GridHeadRoot";
 
 export interface GridColumnHeadersProps<TData extends RowData> {
   table: Table<TData>;
-  width?: CSSProperties["width"];
+  width?: string | number;
   height: string | number;
-  headRef: Ref<HTMLDivElement>;
 }
 
 export default function GridColumnHeaders<TData extends RowData>({
   table,
   width,
   height,
-  headRef,
 }: GridColumnHeadersProps<TData>) {
   return (
     <GridHeadRoot
-      ref={headRef}
       style={{
         width,
         minHeight: height,
@@ -29,7 +26,7 @@ export default function GridColumnHeaders<TData extends RowData>({
       {table.getHeaderGroups().map((headerGroup) => (
         <GridRow key={headerGroup.id}>
           {headerGroup.headers.map((header) => (
-            <GridHeaderCell header={header} key={header.id} />
+            <GridHeadCell header={header} key={header.id} />
           ))}
         </GridRow>
       ))}
