+++ src/components/GridHeadCell.tsx
@@ -2,18 +2,22 @@ import { flexRender, Header, RowData } from "@tanstack/react-table";
 import { type TableCellProps, TableSortLabel } from "@mui/material";
 import { GridCell } from "./styled";
 
-export default function GridHeaderCell<TData extends RowData>({
+export default function GridHeadCell<TData extends RowData>({
   header,
   ...props
 }: TableCellProps & { header: Header<TData, unknown> }) {
   const sortDirection = header.column.getIsSorted() || undefined;
+  const width = header.getSize();
 
   return (
     <GridCell
       {...props}
       sortDirection={sortDirection}
       colSpan={header.colSpan}
-      width={header.getSize()}
+      sx={{
+        minWidth: width,
+        maxWidth: width,
+      }}
     >
       <TableSortLabel
         active={!!sortDirection}
