+++ src/components/GridNoRowsOverlay.tsx
@@ -1,6 +1,6 @@
 import { Typography } from "@mui/material";
 import type { ComponentPropsWithoutRef, ReactNode } from "react";
-import { Overlay } from "./styled";
+import Overlay from "./Overlay";
 
 const DEFAULT_NO_ROWS_LABEL = "No rows to display.";
